requirements.txt
bcrypt>=4.0.0
argon2-cffi>=23.1.0
colorama>=0.4.6
tabulate>=0.9.0
flask>=3.0.0
flask-cors>=4.0.0
